+++ config.py
@@ -604,4 +604,4 @@ CONFIG_EXECUTION_ORDER = (
     "config_debug_layer",
 )
 
-PLUGIN_VERSION = "2.8.0"
+PLUGIN_VERSION = "2.8.1"